fix: flaky ms edge integ test
